Moved old Tree builder code to obsolete folder.
